feat: implement correct floodFill
